Filter gallery by valid lists pulled from the validation service.
